QUR-5 Изменения в соответствии с комментариями ревьюера
